Fix overflow protection in ft_calloc and type casting in ft_putnbr_fd

+++ ft_putnbr_fd.c
@@ -23,6 +23,6 @@ void	ft_putnbr_fd(int n, int fd)
 		nbr = -nbr;
 	}
 	if (nbr >= 10)
-		ft_putnbr_fd(nbr / 10, fd);
+		ft_putnbr_fd((int)(nbr / 10), fd);
 	ft_putchar_fd((nbr % 10) + '0', fd);
 }
